finished operator. I changed its type to a size_t so it can return bucket.value, also set up an alpha if statment in insert that makes sure that alpha is >= 0.5 if not it will resize

+++ HashTable.h
@@ -57,7 +57,7 @@ private:
 
     optional<int> get(const string& key) const;
 
-    int& operator[](const string& key);
+    size_t& operator[](const string& key);
 
     vector<string> keys() const;
